+++ src/app/views/Bech32Entity.svelte
@@ -5,6 +5,7 @@
   import {warn} from "src/util/logger"
   import Content from "src/partials/Content.svelte"
   import NoteDetail from "src/app/views/NoteDetail.svelte"
+  import NaddrDetail from "src/app/views/NaddrDetail.svelte"
   import PersonDetail from "src/app/views/PersonDetail.svelte"
   import {sampleRelays} from "src/agent/relays"
 
@@ -17,6 +18,7 @@
   onMount(() => {
     try {
       ;({type, data} = nip19.decode(entity) as {type: string; data: any})
+      console.log(data)
       relays = sampleRelays((data.relays || []).map(objOf("url")))
     } catch (e) {
       warn(e)
@@ -31,6 +33,10 @@
 {:else if type === "note"}
   <Content>
     <NoteDetail note={{id: data}} />
+  </Content>
+{:else if type === "naddr"}
+  <Content>
+    <NaddrDetail {...data} />
   </Content>
 {:else if type === "nprofile"}
   <PersonDetail npub={nip19.npubEncode(data.pubkey)} {relays} activeTab="notes" />